% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_central_morph.R
\name{calc_central_morph}
\alias{calc_central_morph}
\title{Mean or Median Central Landmark Configuration}
\usage{
calc_central_morph(array, method = "median")
}
\arguments{
\item{array}{A (landmark x dimension x individual) array of landmark coordinates}

\item{method}{Parameter to define the calculation of either the "mean" or the "median" central configuration}
}
\value{
A matrix containing the mean or median central configuraion
}
\description{
The present function can be used to calculate the central configuration of a set of
landmark coordinates. To comply with robust statistical parameters, this function provides
the option of both the median and mean landmark configuration
}
\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$rotated)

}
