% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_optimal_pc_scores.R
\name{calculate_optimal_pc_scores}
\alias{calculate_optimal_pc_scores}
\title{Calculate the Optimal Number of PC Scores}
\usage{
calculate_optimal_pc_scores(
  data,
  N_permutations = 10000,
  centre = TRUE,
  scale = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the data that
will be used for Principal
Component Analysis}

\item{N_permutations}{An integer defining the number of permutations to be
used in the calculation.}

\item{centre}{A boolean variable specifying whether data is to be centred first}

\item{scale}{A boolean variable specifying whether data is to be scaled first}
}
\value{
A plot presenting the results from the permuted calculation of the
optimal number PC scores
}
\description{
The present function can be used to calculate the optimal number of PC Scores
required for a study
}
\section{Details}{

The calculation of the optimal number of PC scores is performed by calculating the
observed variance explained by each PC with the permuted variance.
}

\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)

# calculate optimal number of PC scores
data <- vector_from_landmarks(GPAshape$rotated)

calculate_optimal_pc_scores(data, N_permutations = 10000)
 
}
