% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lm_graph.R
\name{create_lm_graph}
\alias{create_lm_graph}
\title{Create Landmark Graph}
\usage{
create_lm_graph(target_central, target_edges)
}
\arguments{
\item{target_central}{A (landmark x dimension) matrix containing the central landmark
configuration the user wishes to analyse}

\item{target_edges}{An edge list that defines the graph, using
\code{\link{as_edge_list}}}
}
\value{
An object of class igraph
}
\description{
This function can be used to create a landmark graph for further more detailed
network analyses of graph theory applications using the \code{igraph} library
}
\section{Details}{

The objective of this function is to create an igraph object for more user specific
analyses (that may not be included in the GraphGMM library)
}

\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$rotated)

# compute graph edges
edges <- triangulate2d(central_config)

# extract edge list
edge_list <- as_edge_list(edges)

# create igraph object
lm_graph <- create_lm_graph(central_config, edge_list)

}
\seealso{
\code{igraph} library
}
