% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_resistant_fit.R
\name{generalised_resistant_fit}
\alias{generalised_resistant_fit}
\title{Generalised Resistant Fit (GRF)}
\usage{
generalised_resistant_fit(
  lm_array,
  scale = TRUE,
  robust = FALSE,
  tolerance_limit = 0.005,
  projection = c("orthogonal", "stereographic")
)
}
\arguments{
\item{lm_array}{A (num_landmark x num_dimension x num_individuals) array (or tensor) containing
raw landmark coordinates}

\item{scale}{A boolean value indicating whether analyses should be performed in
shape (scale = TRUE) or form (scale = FALSE) space.}

\item{robust}{A boolean value indicating whether a robust statistics should be used.}

\item{tolerance_limit}{A numeric value defining the tolerance acceptable for
algorithm convergence}

\item{projection}{Either select "orthogonal" or "stereographic" for this parameter.
This defines whether Procrustes coordinates are projected into tanjent space
using an orthogonal or stereographic projection.}
}
\value{
The superimposed Procrustes coordinates
}
\description{
The present function performs a full Generalised Resistant Analysis for the
registration of landmark configurations into a superimposed configuration.
This function by default performs superimposition using translation, rotation,
and scaling (shape analysis), however has also been adapted for analyses without
scaling (form analysis).
}
\section{Details}{

This function calculates a full Generalised Resistant fit on a set of 2D or 3D landmark
coordinates. The final product is a set of projected superimposed coordinates. GRF
was developed as a means of reducing the effect a single landmark may have on the
entire configuration (Siegel and Benson, 1982; Siegel and Pinkerton, 1982;
Rohlf and Slice, 1990; Slice, 1996; Dryden and Walker, 1999). Nevertheless,
this method is computationally expensive,
and in some cases do not guarantee convergence.
}

\section{Note}{


This is only an experimental function, we recommend using traditional GPA.

If this function produces an error, it will be due to a lack of convergence
in the algorithm.
}

\section{Bibliography (and additional references)}{


Siegel, A.F.; Benson, R.H. (1982) A robust comparison of biological shapes
Biometrics. 38(2):341-350.

Siegel, A.F.; Pinkerton, J.R. (1982) Robust comparison of three dimensional shapes
with an application to protein molecule configurations, Technical Report,
Department of Statistics, Princeton University. 224

Rohlf, J.F.; Slice, D.E. (1990) Extension of the Procrustes method for the
optimal superimposition of landmarks, Systematic Biology. 39:40-59

Slice, D.E. (1996) Three-dimensional, generalized resistant fitting and the
comparison of least-squares and resistant fit residuals. In: L.F. Marcus,
M. Corti, A. Loy, G.J.P. Naylor, D.E. Slice (Eds.) Advances in Morphometrics,
New York: Plenum Press. 179-199.

Dryden, I.L.; Walker, G. (1999) Highly Resistant Regression and Object Matching,
Biometrics, 55:820-825.

Claude, J. (2008) Morphometrics with R. The Netherlands: Springer

Courtenay, L.A.; Aramendi, J.; González-Aguilera, D. (In Prep) A Graph
Based Geometric Morphometric approach to the analysis of primate radii:
A new mathematical model for the processing of landmark data.
}

\examples{

library(shapes)
data(apes)

# for issues with computational time we set the tolerance to 1

GRFshape <- generalised_resistant_fit(apes$x,
                                      tolerance_limit = 1)
pca_plot(vector_from_landmarks(GRFshape), apes$group)

}
\author{
Lloyd A. Courtenay
}
