% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPA.R
\name{GPA}
\alias{GPA}
\title{Generalised Procrustes Analysis (GPA)}
\usage{
GPA(
  lm_array,
  scale = TRUE,
  robust = FALSE,
  proc_method = c("optimalLS", "LS"),
  projection = c("orthogonal", "stereographic")
)
}
\arguments{
\item{lm_array}{A (num_landmark x num_dimension x num_individuals) array (or tensor) containing
raw landmark coordinates}

\item{scale}{A boolean value indicating whether analyses should be performed in
shape (scale = TRUE) or form (scale = FALSE) space.}

\item{robust}{A boolean value indicating whether a Generalised Robust Fit (GRF)
should be performed.}

\item{proc_method}{Either select "optimalLS" or "LS" for this parameter. See "Details"
for information on the difference between the two approaches}

\item{projection}{Either select "orthogonal" or "stereographic" for this parameter.
This defines whether Procrustes coordinates are projected into tanjent space
using an orthogonal or stereographic projection. This is only used if the proc_method
is set to "LS".}
}
\value{
\code{coordinates} - The Superimposed Coordinates

\code{centroid_sizes} - A list of centroid sizes for each individual
}
\description{
The present function performs a full Generalised Procrustes Analysis for the
registration of landmark configurations into a superimposed configuration.
This function by default performs superimposition using translation, rotation,
and scaling (shape analysis), however has also been adapted for analyses without
scaling (form analysis). GPA is performed using a Least Square (LS) approach.
}
\section{Details}{

This function calculates a full Generalised Procrustes fit on a set of 2D or 3D landmark
coordinates. The final product is either a set of superimposed coordinates on the Procrustes
(hyper)hemisphere (Slice, 2001), or their projection into tanject space
according to either an orthogonal (default) or a stereographic projection. If proc_method = "optimalLS",
then the coordinates are already provided as projected coordinates.

The Least-Square algorithm employed in the present study has been implemented using
two different techniques. The vanilla "LS" approach employs the classic superimposition
techniques described by Gower (1975), while "optimalLS" employs the use of a more
powerful superimposition algorithm that ensures are more optimal convergence
(Rohlf and Slice, 1990). The "optimalLS" approach also includes some computational
optimisations used by Ian Dryden in the "shapes" library (Dryden and Mardia, 2016).

The robust parameter is used for what Courtenay et al. (In Prep) describe as a
Generalised Robust Fit (GRoF), which uses robsut statistical measures (Höhle and Höhle, 2009)
for the calculation of both the centroid and the central configuration
(used as the reference for LS optimisation).
}

\section{Notes}{


At present GRoF is in early stages of development, and the GPA function at present
works best when using robust = FALSE and proc_method = "optimalLS". All other parameters
with the exception of scale work similiar to the \code{procGPA} function of the
\code{shapes} library.

Optimal results are therefore obtained using:

GPA(x, scale = TRUE, robust = FALSE, proc_method = "optimalLS")

or

GPA(x, scale = FALSE, robust = FALSE, proc_method = "optimalLS")

The GRoF function (x, robust = TRUE) is therefore only meant to be used
for experimental purposes (in this version of GraphGMM)

If users wish to use a more stable version of GPA, we strongly recommend the
\code{procGPA} function of the \code{shapes} library
}

\section{Bibliography (and additional references)}{


Gower, J.C. (1975) Generalized Procrustes Analysis, Psychometrika. 40:33-50

Kendall, D.G. (1984) Shape maniforlds, procrustean metrics, and complex projective
spaces, Bulletin of the London Mathematical Society. 16:81-121

Bookstein, F.L. (1986) Size and shape spaces for landmark data in two dimensions,
Statistical Science. 4(2):181-242

Rohlf, J.F.; Slice, D.E. (1990) Extension of the Procrustes method for the
optimal superimposition of landmarks, Systematic Biology. 39:40-59

Goodall, C. (1991) Procrustes Methods in the Statistical Analysis of Shape,
Journal of the Royal Statistical Society B. 53(2):285-339

Slice, D. (2001) Landmark coordinates aligned by Procrustes Analysis do not
lie in Kendall’s shape space, Systematic Biology. 50(1):141-149.
DOI: 10.1080/10635150119110

Claude, J. (2008) Morphometrics with R. The Netherlands: Springer

Höhle, J.; Höhle, M. (2009) Accuracy assessment of digital evelation models
by means of robust statistical methods, ISPRS Journal of Photogrammetry and
Remote Sensing. 64:398-406

Dryden, I.L.; Mardia, K.V. (2016) Statistical Shape Analysis: With Applications
in R. West Sussex: Wiley

Courtenay, L.A.; Aramendi, J.; González-Aguilera, D. (In Prep) A Graph
Based Geometric Morphometric approach to the analysis of primate radii:
A new mathematical model for the processing of landmark data.
}

\examples{

library(shapes)
data(apes)

# shape analysis

GPAshape <- GPA(apes$x, scale = TRUE)

pca_plot(vector_from_landmarks(GPAshape$coordinates), apes$group)

# form analysis

GPAform <- GPA(apes$x, scale = FALSE)

pca_plot(vector_from_landmarks(GPAform$coordinates), apes$group)

}
\author{
Lloyd A. Courtenay (with elements inspired by Ian Dryden's "shapes" package,
and Claude, 2008)
}
