% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphGMM_support.R
\docType{package}
\name{GraphGMM}
\alias{GraphGMM}
\title{A library for Graph based Geometric Morphometrics}
\description{
An extension to geometric morphometric libraries oriented towards the
integration of Graph Theory into the analysis of landmark data. Graph-based
Geometric Morphometrics is additionally accompanied by other non-linear dimensionality
reduction algorithms.
}
\author{
Lloyd A. Courtenay
}
