% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_pca_biplot.R
\name{kernel_pca_biplot}
\alias{kernel_pca_biplot}
\title{Non-Linear Principal Component Analysis Biplot}
\usage{
kernel_pca_biplot(
  data,
  kernel,
  hyperparam = 0,
  labels = NULL,
  axes = c(1, 2),
  label_colours = NULL,
  centre = TRUE,
  scale = FALSE,
  n_variables = 5,
  arrow_size = 1.5,
  arrow_colour = "#CD5C5C",
  point_size = 1,
  CI_ellipse = FALSE,
  CI_level = 0.95,
  CI_size = 1,
  Chull = FALSE,
  Ch_size = 1,
  legend_position = "topright",
  legend_inset = 0.02,
  main = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} containing the information to plot}

\item{kernel}{A string defining the type of kernel to be used;
"rbf" (Radial Basis Function), "poly" (Polynomial), "laplace"
(Laplace), "gauss" (Gaussian), "cauchy" (Cauchy), "spline" (Spline)}

\item{hyperparam}{A positive numeric value defining the hyperparameter to be
applied to each type of kernel (unavailable for the Spline function, and not
recommended for the Cauchy function)}

\item{labels}{A factorial list containing the sample labels}

\item{axes}{A two-value vector defining the axes to be plotted (e.g. c(1,2))}

\item{label_colours}{A string vector of colours to be used in the plot.}

\item{centre}{A boolean variable specifying whether data is to be centred first}

\item{scale}{A boolean variable specifying whether data is to be scaled first}

\item{n_variables}{An integer defining the number of variables to be included
in the biplot}

\item{arrow_size}{An integer defining the size of the arrows in the biplot}

\item{arrow_colour}{A string defining the colour of biplot arrows}

\item{point_size}{A numeric value defining the siz eof plot points}

\item{CI_ellipse}{A boolean specifying whether Confidence Ellipses are to be
included}

\item{CI_level}{A numeric value between 0 and 1 defining the % of confidence
for CI ellipses}

\item{CI_size}{A numeric value defining the thikness of CI ellipses lines}

\item{Chull}{A boolean specifying whether convex hulls are to be included}

\item{Ch_size}{A numeric value defining the thickness of Convex Hull lines}

\item{legend_position}{A string defining the position of the legend}

\item{legend_inset}{A numeric value defining the shift in position of the legend}

\item{main}{A string specifying the plot title (if included)}
}
\value{
PCA biplot

\code{variable_contribution} - a list of values defining the weight
each variable has on the biplot.
}
\description{
This function provides an extension of \code{\link{pca_biplot}} to create a
non-linear PCA biplot using a selection of kernels.
}
\section{Details}{

This function plots a Non-Linear Princpial Component Analysis biplot using a
prior transformation to the data according to a kernel provided by the user.
}

\examples{
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- GPA(apes$x)
data <- vector_from_landmarks(GPAshape$coordinates)

# plot non-linear pca

kernel_pca_biplot(data, apes$group, kernel = "rbf", hyperparam = 0.001)

}
\seealso{
\code{\link{kernel_cauchy}}, \code{\link{kernel_gaussian}},
\code{\link{kernel_laplace}}, \code{\link{kernel_poly}},
\code{\link{kernel_rbf}}, \code{\link{kernel_spline}}, \code{\link{kernel_pca}},
\code{\link{pca_plot}}, \code{\link{pca_biplot}}
}
\author{
Lloyd A. Courtenay
}
