% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_poly.R
\name{kernel_poly}
\alias{kernel_poly}
\title{The Polynomial Kernel for Non-Linear Transformations}
\usage{
kernel_poly(X, degree = 2)
}
\arguments{
\item{X}{A matrix of data to be transformed}

\item{degree}{The degree hyperparameter defining a scaling degree}
}
\value{
Transformed matrix of values
}
\description{
The Polynomial Kernel for Non-Linear Transformations
}
\section{Bibliography}{

Bishop, C. (2006) Pattern Recognition and Machine Learning. Singapore: Springer.

Cortes, C.; Vapnik, V. (1995) Support Vector Networks, Machine Learning, 20:273-297
DOI: 10.1007/BF00994018
}

\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)
data <- vector_from_landmarks(GPAshape$rotated)

# perform non linear transformation

transformed_data <- kernel_poly(data)

# plot non-linear pca

pca_plot(transformed_data, apes$group)

}
\author{
Lloyd A. Courtenay
}
