% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_rbf.R
\name{kernel_rbf}
\alias{kernel_rbf}
\title{The Radial Basis Function Kernel for Non-Linear Transformations}
\usage{
kernel_rbf(X, gamma = 1)
}
\arguments{
\item{X}{A matrix of data to be transformed}

\item{gamma}{The gamma hyperparameter defining a scaling degree}
}
\value{
Transformed matrix of values
}
\description{
The Radial Basis Function Kernel for Non-Linear Transformations
}
\section{Bibliography}{

Bishop, C. (2006) Pattern Recognition and Machine Learning. Singapore: Springer.

Cortes, C.; Vapnik, V. (1995) Support Vector Networks, Machine Learning, 20:273-297
DOI: 10.1007/BF00994018
}

\examples{
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- GPA(apes$x)
data <- vector_from_landmarks(GPAshape$rotated)

# perform non linear transformation

transformed_data <- kernel_rbf(data)

# plot non-linear pca

pca_plot(transformed_data, apes$group)

}
\author{
Lloyd A. Courtenay
}
