% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_spline.R
\name{kernel_spline}
\alias{kernel_spline}
\title{The Spline Kernel for Non-Linear Transformations}
\usage{
kernel_spline(X)
}
\arguments{
\item{X}{A matrix of data to be transformed}
}
\value{
Transformed matrix of values
}
\description{
The Spline Kernel for Non-Linear Transformations
}
\section{Bibliography}{

Gunn, S. (1998) Support Vector Machines for Classification and Regression,
ISIS Technical Report. http://www.svms.org/tutorials/Gunn1998.pdf
}

\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)
data <- vector_from_landmarks(GPAshape$rotated)

# perform non linear transformation

transformed_data <- kernel_spline(data)

# plot non-linear pca

pca_plot(transformed_data, apes$group)

}
\author{
Lloyd A. Courtenay
}
