% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_graph.R
\name{knn_graph}
\alias{knn_graph}
\title{k-Nearest Neighbour Calculation}
\usage{
knn_graph(central_config, k = 3, radius = NULL)
}
\arguments{
\item{central_config}{A (landmark x dimension) matrix containing the
central configuration that will be used to calculate the graph.}

\item{k}{An integer defining the number of Nearest Neighbours.}

\item{radius}{A numeric value defining the radius around a point
to be considered a neighbour. If a radius is not provided, then all points can
be considered neighbours.}
}
\value{
A list of edges.
}
\description{
A nearest neighbour function to compute the spatial relationships between landmarks
}
\section{Details}{

This function calculates the nearest neighbours of each landmark.
}

\examples{
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- GPA(apes$x)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$coordinates)

# compute graph edges
edges <- knn_graph(central_config)

plot_landmark_graph(central_config, edges)

}
\author{
Lloyd A. Courtenay
}
