% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_modularity.R
\name{landmark_modularity}
\alias{landmark_modularity}
\title{Calculate Landmark Modules}
\usage{
landmark_modularity(
  target_central,
  target_edges,
  dimensions = c("x", "y"),
  point_size = 2,
  Chull = TRUE,
  Ch_size = 1,
  Ch_alpha = 0,
  plot_type = "2D",
  main = NULL
)
}
\arguments{
\item{target_central}{A (landmark x dimension) matrix containing the central landmark
configuration the user wishes to analyse}

\item{target_edges}{An edge list that defines the graph, using
\code{\link{as_edge_list}}}

\item{dimensions}{A two-value vector defining the dimensions to plot in the case
of using a 2D visualisation (e.g. c("x","y"))}

\item{point_size}{A numeric value defining the size of plot points}

\item{Chull}{A boolean specifying whether convex hulls are to be included}

\item{Ch_size}{A numeric value defining the thickness of Convex Hull lines}

\item{Ch_alpha}{A numeric value defining how transparent the convex hulls are}

\item{plot_type}{"2D" or "3D" defining the type of visualisation}

\item{main}{A string specifying the plot title (if included)}
}
\value{
\code{modularity_results} - a table of landmarks and their corresponding
module

\code{visualisation} - a ggplot object visualising the modules (for 2D
visualisation only)
}
\description{
A function used to calculate and plot the landmark modules for a given configuration
}
\section{Details}{

This function uses the Louvian algorithm to calculate neighbourhood structure
(modularity) within the landmark graph.
}

\section{Bibliography}{

Blondel, V.D.; Guillaume, J.L.; Lambiotte, R.; Lefebvre, E. (2008)
Fast unfolding of communities in large networks. Journal of Statistical
Mechanics. arXiv: 0803.0476v2
}

\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)

# separate the samples under analysis

female_gorilla <- separate_samples(GPAshape$rotated,
                                   apes$group,
                                   "gorf")
male_gorilla <- separate_samples(GPAshape$rotated,
                                 apes$group,
                                 "gorm")

female_central <- calc_central_morph(female_gorilla)
male_central <- calc_central_morph(male_gorilla)

# calculate edges

female_edges <- knn_graph(female_central, k = 3, radius = 75)
male_edges <- knn_graph(male_central, k = 3, radius = 75)
female_edges <- as_edge_list(female_edges)
male_edges <- as_edge_list(male_edges)

# Graph configuration statistics

landmark_modularity(female_central, female_edges, Chull = FALSE,
                    point_size = 5)
landmark_modularity(male_central, male_edges, Chull = FALSE,
                    point_size = 5)

}
\seealso{
\code{\link[igraph]{cluster_louvain}} from the Igraph library
}
