% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahlanobis_dist_matrix.R
\name{mahalanobis_dist_matrix}
\alias{mahalanobis_dist_matrix}
\title{Mahalanobis Distances and p-Values}
\usage{
mahalanobis_dist_matrix(
  data,
  labels,
  plot_results = TRUE,
  robust = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A matrix containing the data}

\item{labels}{A factorial list containing the sample labels}

\item{plot_results}{A boolean value defining whether results are to be plotted.
A maximum of 4 groups can be included for results to be plotted.}

\item{robust}{A boolean value defining whether p-values are to be computed
using robust statistical metrics.}

\item{verbose}{A boolean value indicating whether a progress bar should be printed across
screen}
}
\value{
\code{distances} - A distance matrix

\code{p_values} - p-Values computed from the distance matrix
}
\description{
A function used to calculate inter and intra-group mahalanobis distances. p-Values
are computed to calculate statistical differences between samples
}
\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# example 1 ----------------------------------

# load data
data(plethodon)

# Generalized Procrustes Fit
GPAshape <- GPA(plethodon$land)
data <- vector_from_landmarks(GPAshape$coordinates)

# pca

pca <- pca_plot(data, plethodon$species)

pca$pca_plot # visualise pca
pc_scores <- pca$pc_scores # extract pc_scores

# mahalanobis distances

mahalanobis_dist_matrix(pc_scores, plethodon$species)

# example 2 ----------------------------------

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- GPA(apes$x)
data <- vector_from_landmarks(GPAshape$coordinates)

# pca

pca <- pca_plot(data, apes$group)

pca$pca_plot # visualise pca
pc_scores <- pca$pc_scores # extract pc_scores

# mahalanobis distances

mahalanobis_dist_matrix(pc_scores, apes$group, plot_results = FALSE)

}
\seealso{
\code{\link{procD_comparisons}}
}
\author{
Lloyd A. Courtenay
}
