% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{Principal Component Analysis}
\usage{
pca_plot(
  data,
  labels = NULL,
  axes = c(1, 2),
  label_colours = NULL,
  centre = TRUE,
  scale = FALSE,
  point_size = 2,
  CI_ellipse = FALSE,
  CI_level = 0.95,
  CI_size = 1,
  Chull = FALSE,
  Ch_size = 1,
  main = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} containing the information to plot}

\item{labels}{A factorial list containing the sample labels}

\item{axes}{A two-value vector defining the axes to be plotted (e.g. c(1,2))}

\item{label_colours}{A string vector of colours to be used in the plot.}

\item{centre}{A boolean variable specifying whether data is to be centred first}

\item{scale}{A boolean variable specifying whether data is to be scaled first}

\item{point_size}{A numeric value defining the siz eof plot points}

\item{CI_ellipse}{A boolean specifying whether Confidence Ellipses are to be
included}

\item{CI_level}{A numeric value between 0 and 1 defining the % of confidence
for CI ellipses}

\item{CI_size}{A numeric value defining the thikness of CI ellipses lines}

\item{Chull}{A boolean specifying whether convex hulls are to be included}

\item{Ch_size}{A numeric value defining the thickness of Convex Hull lines}

\item{main}{A string specifying the plot title (if included)}
}
\value{
\code{variance} - Explained variance of each PC Score

\code{pc_scores} - The PC scores

\code{pca_plot} - a ggplot object containing the pca plot
}
\description{
This function calculates and plots a PCA.
}
\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)
data <- vector_from_landmarks(GPAshape$rotated)

# plot pca

pca_plot(data, apes$group, CI_ellipse = TRUE)

}
\seealso{
\code{\link{kernel_pca}},
\code{\link{pca_biplot}},
\code{\link{kernel_pca_biplot}}
}
\author{
Lloyd A. Courtenay
}
