% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landmark_graph.R
\name{plot_landmark_graph}
\alias{plot_landmark_graph}
\title{Plot a landmark graph}
\usage{
plot_landmark_graph(
  central_config,
  edges,
  point_size = NULL,
  point_colour = NULL,
  line_size = NULL,
  line_colour = NULL
)
}
\arguments{
\item{central_config}{A (landmark x dimension) matrix containing the
central configuration that will be used to calculate the graph.}

\item{edges}{A matrix defining the edges for the graph.}

\item{point_size}{A numeric value defining the size of the points in the visualisation}

\item{point_colour}{A string defining the colour of points.}

\item{line_size}{A numeric value defining the thickness of edge lines}

\item{line_colour}{A string defining the colour of edges.}
}
\value{
A plot of the landmark graph
}
\description{
The present function can be used to visualise the computed graphs on a given landmark
configuration.
}
\examples{
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- GPA(apes$x)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$coordinates)

# compute graph edges
edges <- knn_graph(central_config)

plot_landmark_graph(central_config, edges)

}
