% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_warp_analysis.R
\name{relative_warp_analysis}
\alias{relative_warp_analysis}
\title{Relative Warp Analysis (RWA)}
\usage{
relative_warp_analysis(
  proc_coords,
  labels = NULL,
  robust = FALSE,
  alpha = 1,
  projection = FALSE,
  proj_scale = FALSE,
  axes = c(1, 2),
  label_colours = NULL,
  centre = TRUE,
  scale = FALSE,
  point_size = 2,
  CI_ellipse = FALSE,
  CI_level = 0.95,
  CI_size = 1,
  Chull = FALSE,
  Ch_size = 1,
  main = NULL
)
}
\arguments{
\item{proc_coords}{A (num_landmark x num_dimension x num_individuals) array (or tensor) containing
raw landmark coordinates}

\item{labels}{A factorial list containing the sample labels for RWA plot}

\item{robust}{A boolean value indicating whether robust statistics should be used.}

\item{alpha}{A non-zero numeric value defining the scaling effect the bending matrix
has on RWA results. alpha = 1 will highlight large scale differences (global), while alpha = -1 will
highlight small scale variations (local).}

\item{projection}{Whether a projection of Procrustes coordinates to tangent space is needed}

\item{proj_scale}{Whether an additional scaling procedure is need in the calculation of
the pole (central configuration) during projection onto tangent space}

\item{axes}{A two-value vector defining the axes to be plotted (e.g. c(1,2))}

\item{label_colours}{A string vector of colours to be used in the plot.}

\item{centre}{A boolean variable specifying whether data is to be centred first}

\item{scale}{A boolean variable specifying whether data is to be scaled first}

\item{point_size}{A numeric value defining the siz eof plot points}

\item{CI_ellipse}{A boolean specifying whether Confidence Ellipses are to be
included}

\item{CI_level}{A numeric value between 0 and 1 defining the % of confidence
for CI ellipses}

\item{CI_size}{A numeric value defining the thikness of CI ellipses lines}

\item{Chull}{A boolean specifying whether convex hulls are to be included}

\item{Ch_size}{A numeric value defining the thickness of Convex Hull lines}

\item{main}{A string specifying the plot title (if included)}
}
\value{
\code{relative_warp_scores} - PC Scores obtained from Relative Warp Analysis

\code{nonaffine_variation_exp} - The percentage of non-affine
variation explained by each relative warp score.

\code{nonaffine_variation} - descriptors of nonaffine variation (the Partial Warp Scores)

\code{uniform_scores} - Uniform Scores

\code{affine_variation} - descriptors of affine variation (the Uniform Component)
}
\description{
This function is used to calculate the bending energy derived from multiple Thin
Plate Spline interpolation functions. This bending energy can then be used to describe
patterns in affine and non-affine variation, while using an alpha value to emphasize
both local and global morphological changes. The eigendecomposition of Partial
Warp Scores is known as Relative Warp Analysis.
}
\section{Details}{


Thin Plate Spline (TPS) interpolation functions are used to describe a smooth function
from one specimen to another. This function can be used to plot visual deformations
between two different shapes, or used to calculate multiple components of morphological
variation. When TPS functions are decomposed into their seperate components,
they can be used to describe a series of affine (uniform) and non-affine (not uniform) deformations
between the two shapes.

Affine transformations can include translation, scaling, and shearing
of the overall configuration, without being localised to any particular region of the
configuration (subset of landmarks). Non-affine transformations, on the other hand,
refer to the expansion, compression or bending of local regions.

Affine transformations are described by Uniform shape components.

Non-affine transformations are described by principal warps

Relative Warp Scores are the eigendecomposition of the weighted principal warps
}

\section{Selection of alpha value}{


Please do not select an alpha value of 0. Scaling partial warp scores to 0
has been found to produce almost identical results to a standard PCA on
coordinate data. See Rohlf (1993, 1996, 1999) for more details.
}

\section{Additional Notes}{


Many R libraries perform RWA using their own configuration of GPA. The present
library allows the user to predefine the superimposition procedure they wish to
use. Thus the user must provide the relative_warp_analysis function with
coordinates already superimposed in their desired feature space.
}

\section{Bibliography (and additional references)}{


Bookstein, F.L. (1989) Principal warps: thin-plate splines and the decomposition
of deformations. IEEE Transactions on Pattern Analysis and Machine Intelligence. 11:567-585

Bookstein, F.L. (1991)  Morphometric tools for landmark data: geometry and
biology. Cambridge: Cambridge University Press.

Rohlf, F.J. (1993) Relative warp analysis and an example of its application to
mosquito wings. In: L.F. Marcus, E. Bello, A. García-Valdecasas (Eds.)
Contributions to Morphometrics. Madrid: Monografías del Museo Nacional de
Ciencias Naturales. 8:131-159

Rohlf, F.J. (1996) Morphometric spaces, shape components, and the effects of
linear transformations. In: L.F. Marcus, M. Corti, A. Loy, G.J.P. Naylor, D.E.
Slice (Eds.) Advances in Morphometrics. The Netherlands: Springer. 117-129

Rohlf, F.J., Loy, A., Corti, M. (1996) Morphometric analysis of old world
talpidae (Mammalia, Insectivora) using partial-warp scores, Systematic Biology.
45(3):344-362

Rohlf, F.J. (1999) Shape statistics: Procrustes superimposition and tangent spaces,
Journal of Classification. 16:197-223

Rohlf, F.J.; Bookstein, F.L. (2003) Computing the uniform component of shape
variation. Systematic Biology. 52(1):66-69
}

\examples{

library(shapes)
data(apes)

# shape analysis

GPAshape <- GPA(apes$x, scale = TRUE)

relative_warp_analysis(GPAshape$coordinates, apes$group,
                       alpha = -1)
relative_warp_analysis(GPAshape$coordinates, apes$group,
                       alpha = 1)

}
\author{
Lloyd A. Courtenay
}
