% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps_visualisation.R
\name{tps_visualisation}
\alias{tps_visualisation}
\title{Thin Plate Spline - Shape Predictor}
\usage{
tps_visualisation(
  gpa_coordinates,
  pca_data,
  pcscore,
  pccoord,
  type,
  edges = NULL,
  reference_shape = NULL
)
}
\arguments{
\item{gpa_coordinates}{A (landmark x dimension x individuals) array (or tensor) containing
all landmark coordinates after Generalized Procrustes Analyses.}

\item{pca_data}{The PC Scores from the Principal Component Analysis}

\item{pcscore}{The PC Score(s) that will be used for shape prediction}

\item{pccoord}{The coordinates used for shape prediction}

\item{type}{The type of visualisation; 'graph' visualisation, 'surface' for 3D
surfaces, and 'warpgrid' for thin plate spline warpgrid visualisation (from geomorph)}

\item{edges}{A edge matrix used for 'graph' type visualisation}

\item{reference_shape}{The central landmark configuration used for 'warpgrid'
visualisation}
}
\value{
A plotted visualisation of the predicted shape

The predicted landmark coordinates
}
\description{
The present function can be used to calculate embeddings for each landmark configuration,
producing a final output of each embedded individual, as well as the similarity matrix.
}
\section{Details}{

This function is the central component of the graphGMM library, used to project landmark
configurations into a new embedded feature space, and calculate the similarity matrix
that represents the structural properties of the entire configuration.
}

\section{Bibliography}{

Bookstein, F.L. (1989) Principal Warps: thin plate spline and the decomposition of
deformations, Transactions on Pattern Analysis in Machine Intelligence.
11(6):567-585
}

\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# 2D example --------------------------------------

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$rotated)

# compute graph edges
edges <- triangulate2d(central_config)

# extract edge list
edge_list <- as_edge_list(edges)

# create graph embeddings
graph_object <- graph_embeddings(GPAshape$rotated, edge_list,
                                 num_convolutions = 2)

pca <- pca_plot(graph_object$similarity_vector, apes$group)

pca$pca_plot # visualise pca_plot

# visualise pc1 extremities

tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = max(pca$pc_scores[,1]), type = "graph",
                  edges = edges)
tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = min(pca$pc_scores[,1]), type = "graph",
                  edges = edges)

# visualise pc2 extremities

tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 2,
                  pccoord = max(pca$pc_scores[,2]), type = "graph",
                  edges = edges)
tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 2,
                  pccoord = min(pca$pc_scores[,2]), type = "graph",
                  edges = edges)

# warpgrid visualisation

tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = max(pca$pc_scores[,1]), type = "warpgrid",
                  reference_shape = central_config)
tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = min(pca$pc_scores[,1]), type = "warpgrid",
                  reference_shape = central_config)

# 3D example --------------------------------------

data(macf.dat)

# Generalized Procrustes Fit
GPAshape <- procGPA(macf.dat)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$rotated)

# compute graph edges
edges <- triangulate3d(central_config)

# extract edge list
edge_list <- as_edge_list(edges)

# create graph embeddings
graph_object <- graph_embeddings(GPAshape$rotated, edge_list,
                                 num_convolutions = 2)

pca <- pca_plot(graph_object$similarity_vector)

pca$pca_plot # visualise pca_plot

# visualise pc1 extremities

tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = max(pca$pc_scores[,1]), type = "surface")
tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = min(pca$pc_scores[,1]), type = "surface")

tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = max(pca$pc_scores[,1]), type = "graph",
                  edges = edges)
tps_visualisation(GPAshape$rotated, pca$pc_scores, pcscore = 1,
                  pccoord = min(pca$pc_scores[,1]), type = "graph",
                  edges = edges)

}
\seealso{
\code{\link[geomorph]{shape.predictor}} and \code{\link[geomorph]{plotRefToTarget}}
from the \code{geomorph} library
}
