% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate2d.R
\name{triangulate2d}
\alias{triangulate2d}
\title{2D Delauney Triangulation}
\usage{
triangulate2d(central_config)
}
\arguments{
\item{central_config}{A (landmark x dimension) matrix containing the
central configuration that will be used to calculate the graph.}
}
\value{
A list of edges.
}
\description{
A delauney triangulation function to compute the spatial relationships between landmarks
}
\section{Bibliography}{

Delaunay, B. (1934) Sur la sphère vide.
Bulletin l’Academie de Sciences l’URSS Classe des Sciences Mathematiques et Naturelles
6:793–800.
}

\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$rotated)

# compute graph edges
edges <- triangulate2d(central_config)

plot_landmark_graph(central_config, edges)

}
\seealso{
\code{\link[RTriangle]{triangulate}} of the \code{RTriangle} library
}
