% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate3d.R
\name{triangulate3d}
\alias{triangulate3d}
\title{3D Ball Pivoting Triangulation}
\usage{
triangulate3d(
  central_config,
  pivot_ball_radius = 0,
  pivot_ball_clustering = 0.2,
  pivot_ball_angle = pi/2
)
}
\arguments{
\item{central_config}{A (landmark x dimension) matrix containing the
central configuration that will be used to calculate the graph.}

\item{pivot_ball_radius}{The radius of the ball rolling over the 3D landmarks}

\item{pivot_ball_clustering}{The clustering radius of the ball.}

\item{pivot_ball_angle}{The angle threshold of the ball.}
}
\value{
A list of edges.
}
\description{
A ball-pivoting triangulation function to compute the spatial relationships between landmarks
}
\section{Note}{

The variables clustering and angle for the pivot ball can be used to avoid
the creation of too small triangles.
}

\section{Bibliography}{

Bernardini, F.; Mittleman, J.; Rushmeier, H.; Silva, C.; Taubin, G. (1999)
The ball-pivoting algorithm for surface reconstruction.
IEEE Transaction on Visualization and Computer Graphics, 5(4):349-359
}

\examples{
library(shapes)
library(GraphGMM)

# load data
data(macf.dat)

# Generalized Procrustes Fit
GPAshape <- GPA(macf.dat)

# calculate central configuration
central_config <- calc_central_morph(GPAshape$rotated)

# compute graph edges
edges <- triangulate3d(central_config)

plot_landmark_graph(central_config, edges)

}
\seealso{
\code{\link[Rvcg]{vcgBallPivoting}} of the \code{Rvcg} library
}
