% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_from_landmarks.R
\name{vector_from_landmarks}
\alias{vector_from_landmarks}
\title{Miscellaneous: A function to convert a landmark array into a vector}
\usage{
vector_from_landmarks(landmark_array)
}
\arguments{
\item{landmark_array}{A (landmark x dimension x individual) array containing the landmark
data}
}
\value{
A (individual x (landmark * dimension)) matrix of landmarks
}
\description{
This function can be used to convert a landmark array into a vector for dimensionality
reduction (PCA, non-linear PCA, or tSNE)
}
\examples{
library(geomorph)
library(shapes)
library(GraphGMM)

# load data
data(apes)

# Generalized Procrustes Fit
GPAshape <- procGPA(apes$x)

data <- vector_from_landmarks(GPAshape$rotated)

# plot non-linear pca

pca_plot(data, apes$group)

}
\seealso{
\code{\link{kernel_pca}}, \code{\link{kernel_pca_biplot}},
\code{\link{pca_plot}}, \code{\link{pca_biplot}}
}
