% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_morphologika_file.R
\name{write_morphologika_file}
\alias{write_morphologika_file}
\title{Miscellaneous: Convert an array of landmarks into a morphologika file}
\usage{
write_morphologika_file(file_name, landmarks, labels)
}
\arguments{
\item{file_name}{A string specifying the name of the file that will be produced}

\item{landmarks}{A (landmark x dimension x individual)  or (landmark x dimension)
array containing landmark data}

\item{labels}{A factorial list specifying the label that corresponds to each individual}
}
\value{
Produces a morphologika file.
}
\description{
A miscellaneous function that can be used to format landmarks into a morphologika
file
}
\examples{
library(shapes)
library(GraphGMM)

# load data
data(macf.dat)
data(macm.dat)

library(abind)

mac.dat <- abind(macf.dat, macm.dat, along = 3)

f_list <- rep("F", dim(macf.dat)[3])
m_list <- rep("M", dim(macm.dat)[3])
label_list <- as.factor(c(f_list, m_list))

write_morphologika_file("mac_data", mac.dat,
                        label_list)

}
\author{
Lloyd A. Courtenay
}
